update function location
