+++ client/src/views/TimesheetView.vue
@@ -6,6 +6,8 @@
           <TimesheetListDisplay
             :updateState="updateState"
             :timesheetListDisplayActions="timesheetListDisplayActions"
+            :fetchData="getUserTimesheets"
+            :userTimesheets="userTimesheets"
           />
           <v-container 
             v-if="!isTimesheetListLoading"
@@ -36,6 +38,7 @@
 </template>
 
 <script setup lang="ts">
+import axios from 'axios'
 import { ref } from 'vue'
 import { storeToRefs } from 'pinia'
 
@@ -43,7 +46,7 @@ import type { TimesheetStateTypes } from '../stores/useDataStore'
 import { useHandleTimesheetDisplay } from '../stores/useDataStore'
 import { useLoadingScreen } from '../stores/useUserInterfaceStore'
 import type { Timesheet } from '../stores/types'
-
+import { useGoogleUserData } from '../stores/useDataStore'
 
 import EditTimesheet from '../components/EditTimesheet.vue'
 import TimesheetListDisplay from '../components/TimesheetListDisplay.vue'
@@ -52,6 +55,8 @@ import IsUserLoggedInWrapper from '../components/IsUserLoggedInWrapper.vue'
 const useLoadingScreenStore = useLoadingScreen()
 const { isTimesheetListLoading } = storeToRefs(useLoadingScreenStore)
 const { resetTimesheetDisplay, setTimesheetDisplayStatus, setCurrentTimesheet } = useHandleTimesheetDisplay()
+const { id, isUserLoggedIn } = useGoogleUserData()
+const { setLoadingState } = useLoadingScreen()
 
 const state = ref<TimesheetStateTypes>('allTimesheets')
 
@@ -66,6 +71,8 @@ const handleAddNewTimesheet = () => {
   return
 }
 
+const userTimesheets = ref<Timesheet[]>([])
+
 const timesheetListDisplayActions = ref({
   editTimesheet: (item: Timesheet) => {
     setTimesheetDisplayStatus('edit')
@@ -76,6 +83,35 @@ const timesheetListDisplayActions = ref({
     setTimesheetDisplayStatus('view')
     setCurrentTimesheet(item.timesheetid)
     updateState('editTimesheet')
+  },
+  deleteTimesheet: async (item: Timesheet) => {
+    try {
+      const response = await axios.delete(`/api/timesheets/${item.timesheetid}`)
+
+      if (response.status === 200) {
+        const updatedTimesheets = userTimesheets.value.filter(timesheet => timesheet.timesheetid !== item.timesheetid)
+        userTimesheets.value = updatedTimesheets
+      } else {
+        console.error('Failed to delete timesheet:', response.data.error)
+      }
+    } catch (error) {
+      console.error('Error deleting timesheet:', error)
+    }
   }
 })
+
+const getUserTimesheets = () => {
+  if (!isUserLoggedIn()) return
+
+  setLoadingState('isTimesheetListLoading', true)
+  axios.get(`/api/timesheets/user/${id}`)
+    .then(response => {
+      const { data } = response
+      userTimesheets.value = data.reverse()
+      setLoadingState('isTimesheetListLoading', false)
+    })
+    .catch(error => {
+      console.error('Error fetching data:', error.message)
+    })
+}
 </script>
